Stable evaluation of Gaussian radial basis functions using Hermite polynomials

Stable evaluation of Gaussian radial basis functions using Hermite polynomials

HermiteGF.interpolate_1D — Method.

 interpolate_1D(function_name, nodes_types, epsilon, N, Ne, interpolation_type, gamma)

Computes an interpolant of a 1D function via Hermite-tensor method.

Arguments:

functionname : the name of the tested function. Optionsavailable: "f_3"

nodes_types : an array of two strings with the spacing of nodes in corresponding directions. Options available: :Uniform, :Chebyshev

epsilon : a vector of the values of shape parameters in corresponding directions. The values cannot be larger than 1!

N : a vector with the numbers of basis functions in corresponding directions.

Ne : a vector with the numbers of evaluation points in corresponding directions.

interpolation_type - a string with the name of preferred basis. Options available: :Hermite, :Radial

gamma - a factor allowing to scale the domain and the function accordingly.

© Anna Yurova, 2017

source

HermiteGF.interpolate_2D — Method.

interpolate_2D(function_name, nodes_types, epsilon, N, Ne, interpolation_type, gamma)

Computes an interpolant of a 2D function via Hermite-tensor method.

Arguments:

functionname : the name of the tested function. Optionsavailable: "f_3"

nodes_types : an array of two strings with the spacing of nodes in corresponding directions. Options available: "Uniform", "Chebyshev"

epsilon : a vector of the values of shape parameters in corresponding directions. The values cannot be larger than 1!

N : a vector with the numbers of basis functions in corresponding directions.

Ne : a vector with the numbers of evaluation points in corresponding directions.

interpolation_type : a string with the name of preferred basis. Options available: "Hermite", "Radial"

gamma - a factor allowing to scale the domain and the function accordingly.

© Anna Yurova, 2017

source

HermiteGF.evaluate_hermite — Method.

This function returns the matrix He of the values of the HermiteGF basis functions. The computation is done via three term recurrence for Hermite functions with an argument gamma*x and then appropriate exponential scaling.

More details can be found in Section 5.1 of the paper

STABLE EVALUATION OF GAUSSIAN RADIAL BASIS FUNCTIONS USING HERMITE POLYNOMIALS by Anna Yurova and Katharina Kormann.

This function should be used both for computation of the collocation and evaluation matrices.

source

HermiteGF.evaluate_radial — Method.

Evaluate radial basis

source

HermiteGF.evaluate_s_1D — Method.

Extract the number of collocation and evaluation points

Computing s = X_all * vec(F) (see Sec. 4.1 of the paper)

source

HermiteGF.evaluate_s_2D — Method.

Extract the number of collocation and evaluation points in each dimension

Computing s = (Y_all ⊗ X_all) vec(F) (see Sec. 4.1 of the paper)

source

HermiteGF.evaluate_s_3D — Method.

Extract the number of collocation and evaluation points in each dimension Computing s = (Z_all ⊗ Y_all ⊗ X_all) vec(F) (see Sec. 4.1 of the paper)

source

HermiteGF.evaluate_s_4D — Method.

In this file the functions for the serial computing of the interpolated values 's' in 1-4D are implemented. Here the tensor representation of 's' is used (see Sec. 4.1 of the paper).

Computing s = (Wall ⊗ Zall ⊗ Yall ⊗ Xall) vec(F) (see Sec. 4.1 of the paper)

source

HermiteGF.generate_chebyshev_nodes — Method.

Generate chebyshev nodes on the interval [xmin xmax]

source

HermiteGF.generate_uniform_nodes — Method.

Generate uniform nodes on the interval [xmin xmax]

source

HermiteGF.interpolate_3D — Method.

interpolate_3D(function_name, nodes_types, epsilon, N, Ne, interpolation_type, gamma)

Computes an interpolant of a 3D function via Hermite-tensor method.

Arguments:

functionname : the name of the tested function. Optionsavailable: "f_3"

nodes_types : an array of three strings with the spacing of nodes in corresponding directions. Options available: "Uniform", "Chebyshev"

epsilon : a vector of the values of shape parameters in corresponding directions. The values cannot be larger than 1!

N : a vector with the numbers of basis functions in corresponding directions.

Ne : a vector with the numbers of evaluation points in corresponding directions.

interpolation_type : a string with the name of preferred basis. Options available: :Hermite, :Radial

gamma : a factor allowing to scale the domain and the function accordingly.

© Anna Yurova, 2017

source

HermiteGF.interpolate_4D — Method.

interpolate_4D(function_name, nodes_types, epsilon, N, Ne, interpolation_type, gamma)

Computes an interpolant of a 4D function via Hermite-tensor method.

Arguments:

functionname : the name of the tested function. Optionsavailable: "f_3"

nodes_types : an array of three strings with the spacing of nodes in corresponding directions. Options available: :Uniform, :Chebyshev

epsilon : a vector of the values of shape parameters in corresponding directions. The values cannot be larger than 1!

N : a vector with the numbers of basis functions in corresponding directions.

Ne : a vector with the numbers of evaluation points in corresponding directions.

interpolation_type : a string with the name of preferred basis. Options available: :Hermite, :Radial

gamma - a factor allowing to scale the domain and the function accordingly.

© Anna Yurova, 2017

source

HermiteGF.interpolate_5D — Method.

interpolate_5D(function_name, nodes_types, epsilon, N, Ne, interpolation_type, gamma)

Computes an interpolant of a 5D function via Hermite-tensor method.

Arguments:

functionname : the name of the tested function. Optionsavailable: "f_3"

nodes_types : an array of three strings with the spacing of nodes in corresponding directions. Options available: :Uniform, :Chebyshev

epsilon : a vector of the values of shape parameters in corresponding directions. The values cannot be larger than 1!

N : a vector with the numbers of basis functions in corresponding directions.

Ne : a vector with the numbers of evaluation points in corresponding directions.

interpolation_type : a string with the name of preferred basis. Options available: :Hermite, :Radial

gamma - a factor allowing to scale the domain and the function accordingly.

© Anna Yurova, 2017

source

HermiteGF.trapz_1D — Method.

Trapezoidal rule for the uniform grid for 1-5D

source

interpolation_1D
interpolation_2D
interpolation_3D
interpolation_4D
interpolation_5D