Documentation

HermiteGF.jl Documentation

Stable evaluation of Gaussian radial basis functions using Hermite polynomials

Types

HermiteGF.Chebyshev — Type.

Chebyshev( xmin, xmax, nx )

Chebyshev nodes

source

HermiteGF.Hermite — Type.

Hermite( nodes::NodesType, epsilon:Real, gamma:Real )

Hermite inteprolation

source

HermiteGF.Hermite — Method.

Hermite( xe )

This function returns the matrix He of the values of the HermiteGF basis functions. The computation is done via three term recurrence for Hermite functions with an argument gamma*x and then appropriate exponential scaling.

More details can be found in Section 5.1 of the paper

STABLE EVALUATION OF GAUSSIAN RADIAL BASIS FUNCTIONS USING HERMITE POLYNOMIALS by Anna Yurova and Katharina Kormann.

source

HermiteGF.Radial — Type.

Radial( nodes::NodesType, epsilon, gamma )

Radial interpolation

source

HermiteGF.Radial — Method.

Radial( xe )

returns evaluated function

source

HermiteGF.Uniform — Type.

Uniform( nx, xmin, xmax )

Uniform nodes

source

HermiteGF.InterpolationType — Type.

Interpolation type (Hermite or Radial)

source

HermiteGF.NodesType — Type.

Node positions (Uniform or Chebyshev)

source