Documentation

HermiteGF.jl Documentation

Stable evaluation of Gaussian radial basis functions using Hermite polynomials

HermiteGF.interpolate — Method.

 interpolate( interp :: InterpolationType, 
              f  :: Array{Float64,1},
              xe :: Array{Float64,1})

Computes interpolation of a 1D function via Hermite-tensor method.

Arguments:

interp : inteprolation type (Hermite or Radial)

f : 1d array containing values at interpolation nodes (Chebyshev or Uniform)

xe : vector of evaluation points

© Anna Yurova, 2017

source

HermiteGF.interpolate — Method.

 interpolate( interp_x :: InterpolationType, 
              interp_y :: InterpolationType, 
              f  :: Array{Float64,2},
              xe :: Array{Float64,1},
              ye :: Array{Float64,1})

Computes interpolation of a 2D function via Hermite-tensor method.

Arguments:

interp : inteprolation type (Hermite or Radial)

f : 2d array containing values at interpolation nodes (Chebyshev or Uniform)

xe,ye : vector of evaluation points

source

HermiteGF.interpolate — Method.

 interpolate( interp_x :: InterpolationType, 
              interp_y :: InterpolationType, 
              interp_z :: InterpolationType, 
              f  :: Array{Float64,3},
              xe :: Array{Float64,1},
              ye :: Array{Float64,1},
              ze :: Array{Float64,1})

Computes interpolation of a 3D function via Hermite-tensor method.

Arguments:

interp_x : inteprolation type (Hermite or Radial)

interp_y : inteprolation type (Hermite or Radial)

interp_z : inteprolation type (Hermite or Radial)

f : 3d array containing values at interpolation nodes (Chebyshev or Uniform)

Computing s = ( z ⊗ y ⊗ x) vec(f) (see Sec. 4.1 of the paper)

source

HermiteGF.interpolate — Method.

 interpolate( interp_x :: InterpolationType, 
              interp_y :: InterpolationType, 
              interp_z :: InterpolationType, 
              interp_v :: InterpolationType, 
              f  :: Array{Float64,4},
              xe :: Array{Float64,1},
              ye :: Array{Float64,1},
              ze :: Array{Float64,1},
              ve :: Array{Float64,1})

Computes interpolation of a 4D function via Hermite-tensor method.

Arguments:

interp_x : inteprolation type (Hermite or Radial)

interp_y : inteprolation type (Hermite or Radial)

interp_z : inteprolation type (Hermite or Radial)

interp_v : inteprolation type (Hermite or Radial)

f : 4d array containing values at interpolation nodes (Chebyshev or Uniform)

Computing s = (v ⊗ z ⊗ y ⊗ x) vec(f) (see Sec. 4.1 of the paper)

source

HermiteGF.interpolate — Method.

 interpolate( interp_x :: InterpolationType, 
              interp_y :: InterpolationType, 
              interp_z :: InterpolationType, 
              interp_v :: InterpolationType, 
              interp_w :: InterpolationType, 
              f  :: Array{Float64,4},
              xe :: Array{Float64,1},
              ye :: Array{Float64,1},
              ze :: Array{Float64,1},
              ve :: Array{Float64,1},
              we :: Array{Float64,1})

Computes interpolation of a 4D function via Hermite-tensor method.

Arguments:

interp_x : interpolation type (Hermite or Radial)

interp_y : interpolation type (Hermite or Radial)

interp_z : interpolation type (Hermite or Radial)

interp_v : interpolation type (Hermite or Radial)

interp_w : interpolation type (Hermite or Radial)

f : 5d arrray containing values at interpolation nodes (Chebyshev or Uniform)

Extract the number of collocation and evaluation points in each dimension Computing s = (w ⊗ v ⊗ z ⊗ y ⊗ x) vec(f) (see Sec. 4.1 of the paper)

source

HermiteGF.Chebyshev — Type.

Chebyshev( nx, xmin, xmax )

Chebyshev nodes

source

HermiteGF.Hermite — Type.

Hermite( nodes::NodesType, epsilon:Real, gamma:Real )

Hermite inteprolation

source

HermiteGF.Hermite — Method.

Hermite( xe )

This function returns the matrix He of the values of the HermiteGF basis functions. The computation is done via three term recurrence for Hermite functions with an argument gamma*x and then appropriate exponential scaling.

More details can be found in Section 5.1 of the paper

STABLE EVALUATION OF GAUSSIAN RADIAL BASIS FUNCTIONS USING HERMITE POLYNOMIALS by Anna Yurova and Katharina Kormann.

source

HermiteGF.Radial — Type.

Radial( nodes::NodesType, epsilon, gamma )

Radial interpolation

source

HermiteGF.Radial — Method.

Radial( xe )

returns evaluated function

source

HermiteGF.Uniform — Type.

Uniform( nx, xmin, xmax )

Uniform nodes

source

HermiteGF.evaluate_hermite — Method.

evaluate_hermite(xk, n, epsilon, gamma)

This function returns the matrix He of the values of the HermiteGF basis functions. The computation is done via three term recurrence for Hermite functions with an argument gamma*x and then appropriate exponential scaling.

More details can be found in Section 5.1 of the paper

STABLE EVALUATION OF GAUSSIAN RADIAL BASIS FUNCTIONS USING HERMITE POLYNOMIALS by Anna Yurova and Katharina Kormann.

This function should be used both for computation of the collocation and evaluation matrices.

source

HermiteGF.InterpolationType — Type.

Interpolation type (Hermite or Radial)

source

HermiteGF.NodesType — Type.

Node positions (Uniform or Chebyshev)

source